+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.4 - 2025-06-09
+
+- Fixed `unsubscribe` controller logic by adding missing `Controller` import
+- Fixed documentation about which Laravel model to declare when using model specific notifications
+- Added documentation about what basic functionality the controller provides
+- Reorganised the readme a bit so that unsubscribe is nearer the end
+- Added more documentation for protected and public variables in the base class
+
 ## v0.3 - 2025-05-31
 
 ### Changed